% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_table_functions.R
\name{build_grid_lt}
\alias{build_grid_lt}
\title{Build link table for grid tile.}
\usage{
build_grid_lt(polys, poly_grid, shape_raster)
}
\arguments{
\item{polys}{(sf) shapes to build link table over}

\item{poly_grid}{(sf) polygon to build pixel polygons in}

\item{shape_raster}{(raster::raster) raster that contains shapes for polygonizing}
}
\value{
(data.table) link table for grid tile or NULL if no applicable geometries remain
}
\description{
Build link table for grid tile.
}
