% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_table_functions.R
\name{clean_geometries}
\alias{clean_geometries}
\title{Clean and extract geometries if necessary.}
\usage{
clean_geometries(sfo)
}
\arguments{
\item{sfo}{(sf) sf object with geometries to extract}
}
\value{
(sf) sf object with geometries extracted
}
\description{
In some scenarios (such as polygons being split by intersection),
GEOMETRYCOLLECTIONs are returned. They do not play well with 
some libraries, so we need to extract them into polygons.
}
\details{
We can also safely remove shapes with no area, such as points
and lines.
}
