% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_table_functions.R
\name{factorize_raster_dim}
\alias{factorize_raster_dim}
\title{Find common factors of N and raster dimensions.}
\usage{
factorize_raster_dim(N, factor_raster)
}
\arguments{
\item{N}{(integer) target number to factor}

\item{factor_raster}{(raster::raster) raster for dimensions to factor}
}
\value{
(numeric) common factors for N and factor_raster
}
\description{
This allows clean partitioning of a raster N times. Will
return NA if no common factorization is possible.
}
\details{
This function will infer a preference of x to y
depending on the dimensions of the raster. So if 
x > y, x will get the larger factor.
}
