% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_table_functions.R
\name{finalize_lt}
\alias{finalize_lt}
\title{Finalize link table generation.}
\usage{
finalize_lt(lt)
}
\arguments{
\item{lt}{(data.table) link table to finish}
}
\value{
(data.table) final link table
}
\description{
Get area fractions, total_area, etc.
}
