% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_table_functions.R
\name{find_Ns}
\alias{find_Ns}
\title{Find all possible partition values over a raster.}
\usage{
find_Ns(factor_raster)
}
\arguments{
\item{factor_raster}{(raster::raster) raster for dimensions to factor}
}
\value{
(numeric) possible partition values
}
\description{
Find all possible partition values over a raster.
}
