% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{LinkTable}
\alias{LinkTable}
\title{R6 class for link table}
\description{
description
}
\details{
details
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{lt}}{link table}

\item{\code{id_raster}}{id raster}

\item{\code{pixel_owners}}{pixel owner data.table}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{LinkTable$new()}}
\item \href{#method-get_pixel_ownership}{\code{LinkTable$get_pixel_ownership()}}
\item \href{#method-unique}{\code{LinkTable$unique()}}
\item \href{#method-verify_raster_coverage}{\code{LinkTable$verify_raster_coverage()}}
\item \href{#method-to_raster}{\code{LinkTable$to_raster()}}
\item \href{#method-to_mask}{\code{LinkTable$to_mask()}}
\item \href{#method-print}{\code{LinkTable$print()}}
\item \href{#method-clone}{\code{LinkTable$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
We need the link table and id raster for reference during
raster operations.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LinkTable$new(lt, id_raster)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lt}}{data.table - link table}

\item{\code{id_raster}}{id raster}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_pixel_ownership"></a>}}
\subsection{Method \code{get_pixel_ownership()}}{
Advance through admin levels, geting the pixel owner
then pruning. This ensures, for example, when getting the
ADM1 pixel owner that the ADM1 is within the ADM0 owner. 
Ownership is based on nested aggregated area fraction.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LinkTable$get_pixel_ownership()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
(data.table) pixel with owner at each level
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-unique"></a>}}
\subsection{Method \code{unique()}}{
This function assumes that there are "OWNER_"
fields corresponding to each ADM level. If any of the 
assigned owners are missing from the link table,
the pixel is dropped.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LinkTable$unique()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
(data.table) with one row per pixel
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-verify_raster_coverage"></a>}}
\subsection{Method \code{verify_raster_coverage()}}{
Find missing pixels in a raster based 
on values in the link table.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LinkTable$verify_raster_coverage(verify_raster)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{verify_raster}}{raster to compare with link table}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
(list) missing: link table with entries for missing pixels,
               extra: vector of extra pixel_ids that are not in the link table
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-to_raster"></a>}}
\subsection{Method \code{to_raster()}}{
Link table rows are converted to raster cells
based on pixel_id. Before conversion, only unique
pixels are kept based on nested greatest area ownership.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LinkTable$to_raster(
  template_raster = NULL,
  field = "ADM0_CODE",
  trim_result = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{template_raster}}{raster to crop output to}

\item{\code{field}}{field to populate raster cells}

\item{\code{trim_result}}{trim output raster?}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
(raster::raster)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-to_mask"></a>}}
\subsection{Method \code{to_mask()}}{
Each value in the raster is converted to 1.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LinkTable$to_mask()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
(raster)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
Overrides print() on this object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LinkTable$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LinkTable$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
