% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_table_functions.R
\name{partition_lt_grid}
\alias{partition_lt_grid}
\title{Partition polys into grid.}
\usage{
partition_lt_grid(N, polys, shape_raster, remove_empty = TRUE)
}
\arguments{
\item{N}{(integer) number of pieces to break raster into}

\item{polys}{(sf) polys to divide}

\item{shape_raster}{(raster::raster) raster to align grid with}

\item{remove_empty}{(boolean) remove empty grid pieces?}
}
\value{
(list) grid tiles and their polygon intersections
}
\description{
Partition polys into grid.
}
