% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{[.LinkTable}
\alias{[.LinkTable}
\title{Allow data.table functions over LinkTable.}
\usage{
\method{[}{LinkTable}(lt_object, ...)
}
\arguments{
\item{lt_object}{(LinkTable) link table to be treated as data.table}

\item{...}{arguments to be passed to data.table bracket function}
}
\value{
(LinkTable) new LinkTable object with data.table modifications
}
\description{
Allow data.table functions over LinkTable.
}
