% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_link_table.R
\name{validate_area}
\alias{validate_area}
\title{Validate link table areas.}
\usage{
validate_area(lt, polys, fail_threshold = 1e-05)
}
\arguments{
\item{lt}{(date.table) link table}

\item{polys}{(sf) shapes used to generate link table}

\item{fail_threshold}{(float) largest allowable ratio of area difference to area}
}
\value{
(logical)
}
\description{
See if computed areas in the link table
align with areas from the shapefile. Allow a threshold for rounding
errors/simple features discrepancies.
}
\details{
There is some approximation in area computations, so we need a threshold
to allow area loss/gain.
}
