% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_link_table.R
\name{validate_borders}
\alias{validate_borders}
\title{Validate link table borders.}
\usage{
validate_borders(lt, polys)
}
\arguments{
\item{lt}{(date.table) link table}

\item{polys}{(sf) shapes used to generate link table}
}
\value{
(logical)
}
\description{
Get borders from shapefile and see if
they line up with shared pixels in the link table. Some discrepancy is 
expected due to the fact that adm units that share pixels in the link
may not directly border. ie. They are separated by a distance less than
the pixel polygon size.
}
\details{
To mitigate this, we first compare the link table with the polygon
borders. The link table should not be missing any borders from the
shapefile. Then we make sure the extra borders from the link table
are those that have water in the pixel.
}
