% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_link_table.R
\name{validate_coverage}
\alias{validate_coverage}
\title{Ensure link table covers all admin regions}
\usage{
validate_coverage(lt, polys)
}
\arguments{
\item{lt}{(date.table) link table}

\item{polys}{(sf) shapes used to generate link table}
}
\value{
(logical)
}
\description{
Ensure link table covers all admin regions
}
