% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_link_table.R
\name{validate_duplicates}
\alias{validate_duplicates}
\title{Ensure link table has no duplicates}
\usage{
validate_duplicates(lt)
}
\arguments{
\item{lt}{(date.table) link table}
}
\value{
(logical) no duplicates?
}
\description{
Ensure link table has no duplicates
}
